NIFI-6510 Extract out predictions into their own DTO
