+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-client-dto/src/main/java/org/apache/nifi/web/api/dto/status/ConnectionStatusSnapshotDTO.java
@@ -34,7 +34,7 @@ public class ConnectionStatusSnapshotDTO implements Cloneable {
     private String sourceName;
     private String destinationId;
     private String destinationName;
-    private Boolean predictionsAvailable;
+    private ConnectionStatusPredictionsSnapshotDTO predictions;
     private Integer flowFilesIn = 0;
     private Long bytesIn = 0L;
     private String input;
@@ -48,13 +48,6 @@ public class ConnectionStatusSnapshotDTO implements Cloneable {
     private String queuedCount;
     private Integer percentUseCount;
     private Integer percentUseBytes;
-    private Long predictedMillisUntilCountBackpressure = 0L;
-    private Long predictedMillisUntilBytesBackpressure = 0L;
-    private Integer predictionIntervalSeconds;
-    private Integer predictedCountAtNextInterval = 0;
-    private Long predictedBytesAtNextInterval = 0L;
-    private Integer predictedPercentCount;
-    private Integer predictedPercentBytes;
 
     /* getters / setters */
     /**
@@ -189,15 +182,15 @@ public class ConnectionStatusSnapshotDTO implements Cloneable {
     }
 
     /**
-     * @return indicator showing if predictions available for this connection
+     * @return predictions for this connection
      */
-    @ApiModelProperty("Indicator showing if predictions available for this connection")
-    public Boolean getPredictionsAvailable() {
-        return predictionsAvailable;
+    @ApiModelProperty("Predictions, if available, for this connection (null if not available)")
+    public ConnectionStatusPredictionsSnapshotDTO getPredictions() {
+        return predictions;
     }
 
-    public void setPredictionsAvailable(Boolean predictionsAvailable) {
-        this.predictionsAvailable = predictionsAvailable;
+    public void setPredictions(ConnectionStatusPredictionsSnapshotDTO predictions) {
+        this.predictions = predictions;
     }
 
     /**
@@ -290,69 +283,6 @@ public class ConnectionStatusSnapshotDTO implements Cloneable {
         this.percentUseBytes = percentUseBytes;
     }
 
-    @ApiModelProperty("The predicted number of milliseconds before the connection will have backpressure applied, based on the queued count.")
-    public Long getPredictedMillisUntilCountBackpressure() {
-        return predictedMillisUntilCountBackpressure;
-    }
-
-    public void setPredictedMillisUntilCountBackpressure(Long predictedMillisUntilCountBackpressure) {
-        this.predictedMillisUntilCountBackpressure = predictedMillisUntilCountBackpressure;
-    }
-
-    @ApiModelProperty("The predicted number of milliseconds before the connection will have backpressure applied, based on the total number of bytes in the queue.")
-    public Long getPredictedMillisUntilBytesBackpressure() {
-        return predictedMillisUntilBytesBackpressure;
-    }
-
-    public void setPredictedMillisUntilBytesBackpressure(Long predictedMillisUntilBytesBackpressure) {
-        this.predictedMillisUntilBytesBackpressure = predictedMillisUntilBytesBackpressure;
-    }
-
-    @ApiModelProperty("The predicted number of queued objects at the next configured interval.")
-    public Integer getPredictedCountAtNextInterval() {
-        return predictedCountAtNextInterval;
-    }
-
-    public void setPredictedCountAtNextInterval(Integer predictedCountAtNextInterval) {
-        this.predictedCountAtNextInterval = predictedCountAtNextInterval;
-    }
-
-    @ApiModelProperty("The configured interval (in seconds) for predicting connection queue count and size (and percent usage).")
-    public Integer getPredictionIntervalSeconds() {
-        return predictionIntervalSeconds;
-    }
-
-    public void setPredictionIntervalSeconds(Integer predictionIntervalSeconds) {
-        this.predictionIntervalSeconds = predictionIntervalSeconds;
-    }
-
-    @ApiModelProperty("The predicted total number of bytes in the queue at the next configured interval.")
-    public Long getPredictedBytesAtNextInterval() {
-        return predictedBytesAtNextInterval;
-    }
-
-    public void setPredictedBytesAtNextInterval(Long predictedBytesAtNextInterval) {
-        this.predictedBytesAtNextInterval = predictedBytesAtNextInterval;
-    }
-
-    @ApiModelProperty("Predicted connection percent use regarding queued flow files count and backpressure threshold if configured.")
-    public Integer getPredictedPercentCount() {
-        return predictedPercentCount;
-    }
-
-    public void setPredictedPercentCount(Integer predictedPercentCount) {
-        this.predictedPercentCount = predictedPercentCount;
-    }
-
-    @ApiModelProperty("Predicted connection percent use regarding queued flow files size and backpressure threshold if configured.")
-    public Integer getPredictedPercentBytes() {
-        return predictedPercentBytes;
-    }
-
-    public void setPredictedPercentBytes(Integer predictedPercentBytes) {
-        this.predictedPercentBytes = predictedPercentBytes;
-    }
-
     @Override
     public ConnectionStatusSnapshotDTO clone() {
         final ConnectionStatusSnapshotDTO other = new ConnectionStatusSnapshotDTO();
@@ -363,7 +293,11 @@ public class ConnectionStatusSnapshotDTO implements Cloneable {
         other.setName(getName());
         other.setSourceId(getSourceId());
         other.setSourceName(getSourceName());
-        other.setPredictionsAvailable(getPredictionsAvailable());
+
+        if (predictions != null) {
+            other.setPredictions(predictions.clone());
+        }
+
         other.setFlowFilesIn(getFlowFilesIn());
         other.setBytesIn(getBytesIn());
         other.setInput(getInput());
@@ -377,13 +311,6 @@ public class ConnectionStatusSnapshotDTO implements Cloneable {
         other.setQueuedSize(getQueuedSize());
         other.setPercentUseBytes(getPercentUseBytes());
         other.setPercentUseCount(getPercentUseCount());
-        other.setPredictedMillisUntilCountBackpressure(getPredictedMillisUntilCountBackpressure());
-        other.setPredictedMillisUntilBytesBackpressure(getPredictedMillisUntilBytesBackpressure());
-        other.setPredictionIntervalSeconds(getPredictionIntervalSeconds());
-        other.setPredictedCountAtNextInterval(getPredictedCountAtNextInterval());
-        other.setPredictedBytesAtNextInterval(getPredictedBytesAtNextInterval());
-        other.setPredictedPercentBytes(getPredictedPercentBytes());
-        other.setPredictedPercentCount(getPredictedPercentCount());
 
         return other;
     }
